initial results
In this step I implement 3 classification models on my data to compare the results.

In this process I calculate the confusion matrix and illustrate it to find the different parts of confusion, also calculate the report classification and finding the best test size for models.

Then, I apply model on various folds to find the weaknesses .

I did all these steps for :
1.Logestic Regression
2.naive Bayes
3.KNN

for the next step I should try to improve the models to increase the accuracy.